refactor: resource layout client 제거 및 컴포넌트 분리
