+++ src/app/(providers)/(root)/[workspaceId]/chat/[id]/(resource)/_components/ResourceLayout/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ResourceLayout';
